+++ components/Layout/index.ts
@@ -1,6 +1,6 @@
 import { styled } from 'styles/stitches.config'
 
-export const Content = styled('main', {
+export const Layout = styled('main', {
   display: 'flex',
   flexFlow: 'column',
   alignItems: 'center',
