+++ components/SendTransaction/index.tsx
@@ -1,13 +1,13 @@
 import { ethers } from 'ethers'
 import { useState } from 'react'
 
+import { Box } from 'components/Box'
 import { Button } from 'components/Button'
-import { Heading } from 'components/Heading'
 import { Input } from 'components/Input'
-import { Text } from 'components/Text'
 import { Transaction } from 'types/transaction'
 
-import { Form, TxHistory, TxLine } from './styled'
+import { TxHistory } from './components/TxHistory'
+import { Form } from './styled'
 
 interface IProps {
   ethProvider: ethers.providers.Web3Provider
@@ -45,7 +45,7 @@ export const SendTransaction = ({ ethProvider }: IProps) => {
   }
 
   return (
-    <>
+    <Box>
       {/* eslint-disable-next-line @typescript-eslint/no-misused-promises */}
       <Form onSubmit={handleSubmitForm}>
         <Input
@@ -62,29 +62,7 @@ export const SendTransaction = ({ ethProvider }: IProps) => {
           Send ETH
         </Button>
       </Form>
-      {tsxData && (
-        <TxHistory>
-          <Heading as="h3" size="4" variant="terniary">
-            Your latest transaction:
-          </Heading>
-          <TxLine>
-            <Text>
-              <a
-                href={`https://rinkeby.etherscan.io/tx/${tsxData.hash}`}
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                {tsxData.hash}
-              </a>
-            </Text>
-            <Text role="img" aria-label="transaction status emoji">
-              {txStatus === Transaction.SUCCESS && <>&#9989;</>}
-              {txStatus === Transaction.PENDING && <>&#9203;</>}
-              {txStatus === Transaction.ERROR && <>&#10060;</>} - {txStatus}
-            </Text>
-          </TxLine>
-        </TxHistory>
-      )}
-    </>
+      {tsxData && <TxHistory txData={tsxData} txStatus={txStatus} />}
+    </Box>
   )
 }
