+++ components/SendTransaction/styled.ts
@@ -1,0 +1,6 @@
+import { styled } from 'styles/stitches.config'
+
+export const Form = styled('form', {
+  display: 'flex',
+  gap: '$2',
+})
