+++ constants/appMetaData.ts
